Move sensors into sensor package
