+++ src/main/java/frc/robot/RobotContainer.java
@@ -12,8 +12,8 @@ import frc.robot.drivetrain.DriveBase;
 import frc.robot.drivetrain.DriveIO;
 import frc.robot.drivetrain.DriveIOFalcon;
 import frc.robot.drivetrain.commands.control.XboxControllerDriveControl;
-import frc.robot.drivetrain.gyro.GyroIO;
-import frc.robot.drivetrain.gyro.GyroIOPigeon2;
+import frc.robot.sensors.gyro.GyroIO;
+import frc.robot.sensors.gyro.GyroIOPigeon2;
 import org.talon540.control.XboxController.TalonXboxController;
 
 public class RobotContainer {
