+++ src/main/java/frc/robot/drivetrain/DriveBase.java
@@ -14,8 +14,8 @@ import edu.wpi.first.wpilibj.drive.RobotDriveBase;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.constants.Constants.Drivetrain;
 import frc.robot.drivetrain.DriveIO.DriveNeutralMode;
-import frc.robot.drivetrain.gyro.GyroIO;
-import frc.robot.drivetrain.gyro.GyroIOInputsAutoLogged;
+import frc.robot.sensors.gyro.GyroIO;
+import frc.robot.sensors.gyro.GyroIOInputsAutoLogged;
 import org.littletonrobotics.junction.Logger;
 
 public class DriveBase extends SubsystemBase {
